pyproject.toml
[project]
name = "ecommerce-catalog-enrichment-multi-agent-system-langgraph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "langgraph>=0.6.7",
    "langsmith>=0.4.28",
    "openai>=1.107.3",
    "pytest>=8.4.2",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.20",
    "uvicorn>=0.35.0",
]
